backend/requirements.txt
Django==4.2
djangorestframework==3.15
django-cors-headers==4.0
scikit-learn==1.3
pandas==2.2
numpy==1.26
gunicorn==21.2.0
